fix for host files being given offloading passes
